Update patch set 1
Patch Set 1:

The neverallow rule is enforced for user builds (I tested with an eng build). So, should we add an exception for ueventd or should we adopt the other solution for relabeling the node: unlink the node if it exists and recreate it within correct context.

Patch-set: 1
Label: Verified=0