Update patch set 1
Patch Set 1:

Any update on this patch? Do you think it is too late to trigger the shutdown or just slideshow is inappropriate for this?
I was doubtfully when I made the patch: granting shutdown privileges to an image viewer, is that realy a good idea?
But I then realized that slideshow is mainly used to display dm-verity warning messages.

Patch-set: 1