Update patch set 1
Patch Set 1:

The write denial is occurring because you're trying to add a new file to that directory. You'd get the same behavior if you added the following line to init.rc:

  write /sys/blah asdf

In this case, you're trying to create the file /sys/blah, which requires write access to the /sys directory, generating a denial. /sys/blah isn't a file in sysfs, and it makes no sense to try to create it.

Patch-set: 1