Allow executing update_engine_sideload from recovery.
The recovery flow for A/B devices allows to sideload an OTA downloaded
to a desktop and apply from recovery. This patch allows recovery to run
update_engine_sideload in the "update_engine" context with the same
permissions that the update_engine daemon has in normal operation, which
includes custom boot_control HAL permissions.

This patch also includes some new permissions for update_engine to talk
back to recovery for reporting its progress via a pipe and to read the
sideload OTA file.

Bug: 27178350
Change-Id: I97b301cb2c039fb002e8ebfb23c3599463ced03a