Update patch set 3
Patch Set 3:

(1 comment)

Reusing the update_engine domain seems odd... For instance, update_engine has a bunch of permissions, such as the ability to access files on /data, which should never be granted to a service running from recovery. Or update_engine can talk to priv_app, which again should never occur while running in recovery.

To me, it feels like we should just add the necessary permissions to the recovery domain, and avoid the recovery->update_engine transition entirely. Having two different entrypoints to update_engine doesn't really seem like it buys us that much...

If the goal was just to allow recovery to do anything that the boot_control_hal can do, that's as simple as modifying recovery.te from:

  type recovery, domain, domain_deprecated;

to:

  type recovery, domain, domain_deprecated, boot_control_hal;

The other recovery_only() permissions that you're granting to update_engine are already available in recovery.te, so keeping things in the same domain simplifies the policy quite a bit.

What is it, specifically, that recovery can't do today that update_engine needs to allow for?

Patch-set: 3
Label: Presubmit-Verified=0