Update patch set 3
Patch Set 3:

(1 comment)

Reusing the update_engine domain makes sense because there are a lot of selinux allow rules in update_engine related to update besides just the HAL: there's all the postinstall mounting/relabeling dance in there as well, which has exceptions to neverallow rules elsewhere. If in the future we change how we do that (for which I have some ideas for O to tight up the security), we would need to make sure that the normal update_engine context and whatever else context we use in recovery case are both updated, and for that duplicating the allow rules seems a bad idea.

If the problem is the extra update_engine.te rules that allow access to /data, then I could make a new update_engine_common attribute and a new update_engine_sideload domain so we put the shared stuff in the attribute instead of duplicating the rules. What do you think?

Patch-set: 3