Update patch set 1
Patch Set 1:

> file_contexts parsing doesn't have a concept of userdebug_or_eng.
 > Having said that, I don't think it's needed here. It would just be
 > unnecessary complexity.

 > file_contexts parsing doesn't have a concept of userdebug_or_eng.
 > Having said that, I don't think it's needed here. It would just be
 > unnecessary complexity.

Policy parsing itself doesn't, but the makefile creates the list of used files, and knows about the difference of build type. So we could make this a file_contexts_userdebug and hook it in like the file_contexts_asan.

(Note: I'm fine with not differentiating, just saying it isn't very hard, and we already do it for ASAN)

Patch-set: 1