Update patch set 2
Patch Set 2:

We don't allow dumpstate access to any other application data. But then bluetooth is part of the platform...

Agreed that this is much better than dumping logs to sdcard. Can we create a location for dumpstate to grab bluetooth logs without granting it access to all of the bluetooth app's data?

Patch-set: 2