Update patch set 4
Patch Set 4:

apologies for the -2, but would like an answer before this goes in. Are we granting access to all of the bluetooth's apps data or to logs stored in /data/misc/bluetooth/logs? I was under the impression based on the discussion here that we were doing the later. This CL grants both.

Patch-set: 4