Update patch set 3
Patch Set 3:

Actually, you know what, I'd like Michael to chime in on this one. While it's true that the sysfs path is placed in the default HAL implementation, I don't know if it's a fair assumption that it will be used by all devices. Your concern here is justified!

Michael, is it OK for /sys/devices/virtual/timed_output/vibrator/enable to be considered device-agnostic?

Patch-set: 3