Update patch set 4
Patch Set 4:

These are the type of rules I would like to move out of system_server with the HAIDL work.  If the vibrator HAL can be forced to be in its own process, then it should be running under the corresponding HAL domain.  Could we move this into a section commented with: "Remove when HALs forced into separate process"?  I was told that different HALs could not be in the same process, so we will eventually reaching this goal.

Patch-set: 4
Reviewer: Gerrit User 1026058 <1026058@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0