Update patch set 4
Patch Set 4:

> These are the type of rules I would like to move out of
 > system_server with the HAIDL work.  If the vibrator HAL can be
 > forced to be in its own process, then it should be running under
 > the corresponding HAL domain.  Could we move this into a section
 > commented with: "Remove when HALs forced into separate process"?  I
 > was told that different HALs could not be in the same process, so
 > we will eventually reaching this goal.

In passhthrough mode, the HAL will continue to be a part of system_server. Now, I don't understand why system_server suddenly needs to have a rule to access this vibrator sysfs when earlier it was able to access it just fine.

That said, I will move this to another section as requested.

Patch-set: 4