Update patch set 2
Patch Set 2:

(4 comments)

I grant fd access in most places where ion_device access is granted.  The idea is that, the ion_device is used to map the fd.  It does not make much sense having one without the other.  But it is also possible that ion_device access is not needed in some places, which I do not intend to fix with this CL.

I do not see (or trigger) any issue with ephemeral_app or shell.  I can remove the fd use if you prefer.

Patch-set: 2
Label: Verified=0