Policy split: add export roletype associations.
First version of split policy meant for roles to be unchangeable and
placed the roles file in the private sepolicy section.  The role
association, however, needs to be present for non-platform domains as
well, since with attributization these domains are in the
domain_N domain, rather than domain.  Export the roletype association,
but not the actual role declaration.

Test: Boot device and observe denials.
Bug: 31369363
Change-Id: I855dd2a8e73f196e65b3777e493920485ce5a491