Move sysfs_vibrator HAL system_server policy access into passthrough-macro.
HAL access may be moved to a separate process than the calling code, in which
case policy access rules corresponding to the HAL implementation should be
provided as part of the HAL implementation sepolicy.  This policy should be
removed from the client process if the processes are split, but otherwise remain
if the HAL is in pass-through mode.

Bug: 32243668
Test: Macro'd policy remains if target_build_treble is not set, disappears
otherwise.
Change-Id: I24b7a8bed2c81eab52600bdd21cc6d464639bd5b