Update patch set 4
Patch Set 4:

Now that we've delayed you long enough =/, could you please convert this to the attribute form for HALs?  see: https://android-review.googlesource.com/#/c/303549/

We're going to have to move all of the hal_* domains over, so might as well do it right now instead of having to check in a follow-up change.

Patch-set: 4