Update patch set 5
Patch Set 5:

Daniel, the problem with that is the static vs dynamic versions of this HAL. "boot_control_hal" was here to encapsulate all the selinux policy requirements of the HAL module on a specific device. This module was dynamically used by update_engine/update_verifier, which now just "use" hal_boot; but in recovery we still need to load this HAL statically into the recovery binary, so the selinux permission needs to be in some attribute. We could merge the boot_control_hal into hal_boot, but that means we would need to add hal_boot to recovery which adds a bunch of binder access to the recovery domain that weren't really needed, and that's one of the reasons we split the update_engine_common attribute so recovery can have only what it needs.

Patch-set: 5