Update patch set 5
Patch Set 5:

There isn't a set of read/write permissions that would work for everybody. The recommended way is to use the misc partition, so we would give rw access to that label, but vendors are allowed to store the metadata somewhere else (and the pixel phones do). Besides where to store the A/B metadata, we still need the static library target with the new API and a way to call it. Filed b/32920003

Patch-set: 5
Reviewer: Gerrit User 1077068 <1077068@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0