Update patch set 1
Patch Set 1:

> Remove it and have the code build assuming WITH_DEXPREOPT_PIC=true?
 > If so, that sounds like a much better solution. In fact, I'd love
 > for us to get to the point where we assume WITH_DEXPREOPT=true
 > also, since it allows for security hardening and makes verified
 > boot more effective.

Andreas: let me know if you file a bug to track the removal of WITH_DEXPREOPT_PIC, and I'll abandon this CL and point to that bug. If, in the meantime, you want to land this CL to prevent corner cases from breaking, I'm happy to revert it after WITH_DEXPREOPT_PIC is removed.

Patch-set: 1