Update patch set 2
Patch Set 2:

When you attempt to relabel a file, two permission checks are performed:

1) Are you allowed to "relabelfrom" the type the file currently is?
2) Are you allowed to "relabelto" the type you're trying to label.

In this case, we're trying to relabel /dev/kmsg from "kmsg_device". And in the next step, we're going to try relabeling to "kmsg_device" again. 

The device node is already labeled correctly, and attempting to set it again doesn't make a lot of sense. It's redundant work. 

It would be better to fix the code so that init doesn't attempt to relabel /dev/kmsg twice. Allowing it in policy feels like the wrong fix.

Patch-set: 2
Reviewer: Gerrit User 1003966 <1003966@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0