Update patch set 4
Patch Set 4:

Amit, it might be better to:
- create a hal_radio sepolicy attribute
- have radio inherit from this attribute

in order to be more consistent with other hal implementations. This would also allow other people to implement the radio hal (even if that's currently not planned).

Thoughts, Dan? This is a special case.

Patch-set: 4
Reviewer: Gerrit User 1120458 <1120458@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0