Update patch set 3
Patch Set 3: Code-Review-1

(1 comment)

This remains an draft experiment in creating a logcat domain. Still not ready.

Access to exec'ing logcat is confusing because in some cases we would prefer they use the logpersist domain if they utilize the -f flag, if they plan on storing content in the filesystem.

But in other cases we want to allow them to read their logs unhindered; but alas without the -f flag.

Can we get logcat to transition to logpersist if the -f flag is used?

Patch-set: 3
Label: Code-Review=-1