Update patch set 2
Patch Set 2:

(3 comments)

I vaguely recall that we were just going to load the webview code in the webview zygote, but not actually start execution of the code until we were in the isolated_app domain. This change starts executing the code before we've dropped privileges. I assume this is intentional but I want to double check....

Patch-set: 2
Label: Verified=0