Update patch set 2
Patch Set 2: -Verified -Autosubmit

(3 comments)

Loading the code without executing it isn't really possible; System.loadLibrary() calls JNI_OnLoad, for a start. Plus, we explicitly *do* want to execute as much as we reasonably can in the prefork environment, to save time and maximise page sharing later; we're not doing very much of this yet (pretty much just JNI_OnLoad and the static initialiser for the java provider class) but being able to add more later is desirable.

The idea is supposed to be that the webview_zygote privileges are already as restrictive as isolated_app in terms of the ability to affect the rest of the system.

Patch-set: 2
Label: Verified=0