Update patch set 2
Patch Set 2:

The service is intended to be invoked via `adb shell cmd coverage ...` which is implemented in frameworks/native/cmds/cmd/cmd.cpp. My understanding is that the coverage service needs to be exported for cmd to have access to it. 

See ag/1619472 for context on the service itself.

Patch-set: 2
Label: Verified=0