Update patch set 4
Patch Set 4:

ya, let's just go back to the original patch...

Can we add a comment to the service entry that this is only intended to work on userdebug / eng builds? Then we'll have better info about why this is added and when it's triggered, and could be a future security optimization.

Patch-set: 4
Label: Verified=0