Update patch set 18
Patch Set 18: Code-Review+2

I'm not sure if the remaining rules are needed or not, but they seem reasonable, and we can do further tightening in the future. There's a big advantage to moving this process to it's own domain, so there's a net improvement here even if this domain is lax.

Patch-set: 18