Take two: move MediaProvider into its own domain
Because DownloadProvider runs in the same process as
MediaProvider, it is also inadvertently put into this
domain. To avoid DOA issues, the new domain is very
nearly a superset of priv_app. The domain allows us
to restrict property_sets to only mediaprovider, which
requires them. The domain can be further hardened in the
future.

Test: Automated tests for permissions in Media and Download Providers
Test: Verify MTP works
Test: Verify App downloads work
Test: Verify ringtone selection works
Bug: 33574909
Change-Id: I787843db5ddb91f0cb3dc8e9300eead2c2965061