Update patch set 14
Patch Set 14:

Ok, will add a few neverallows to this and the next CL to tighten it up. It may be too much to ask for to add a read_events_logd rule at this juncture though?

[TL;DR]

Right now _everyone_ is a log reader as app.te adds log reader rights (with the enhancement to read the event-log-tags file in this CL), the only neverallow I could imagine is to block the ability to write to the file (/system/etc/event-log-tags is on a r/o filesystem). The next CL adds some neverallows in private/logd.te (but I see I missed a spot :-( ) as the issue is that we will add /dev/event-log-tags and /data/misc/logd/event-log-tags in the next CL, these files are writeable, and we need to prevent that via DAC and MAC. jeffv@ is worried about anyone being able to read /dev/event-log-tags though once the event log tag service starts. But this one is a groundwork and simplification for now, and the first step to lock this down. What we envisioned in the future is an events log buffer reader right, which is read_logd() plus tags reading, but that may be counter/joined to the effort to create a logcat.te domain (currently under dispute in a separate CL train). Once logcat.te is created, we can make some meaningful neverallow rules and exercise some real control over who can see these files.

Patch-set: 14