Update patch set 15
Patch Set 15: Verified+1

So I add the neverallow rule, but it looks like it may be a labour of love to make sure we catch all that are supposed to access the event-log-tags file(s).

Readers (actually 'printers' to be exact, in order to translate tags to names) and Writers (isloggable checking) of events both need access. Our hope is only _our_ applications and not some device modules or add-on applications (!), as these would make the neverallow rule a PITA*. We have to be on the lookout for more violations on all devices to be sure. But not those writing to the security events buffer (which as a different authority, and bypasses isloggable), or composing SafetyNet event messages (which has specific bypass to isloggable).

Patch-set: 15
Label: Verified=+1