Update patch set 2
Patch Set 2: Code-Review-1

I would prefer we not accept this change. 

It should be up to the data owner to determine how their properties should or shouldn't show up to dumpstate / shell user. Lots of people add properties for different reasons, and we shouldn't presume to know what those properties contain. This is especially true for device-specific properties, where we may not even work with those partners on a day to day basis. IMHO, this is overly permissive and contrary to the principle of least privilege.

Patch-set: 2
Reviewer: Gerrit User 1003966 <1003966@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1