Update patch set 25
Patch Set 25:

I'm having trouble understanding the intent of this change. It sounds like it's about restricting access to a read-only file which by the sounds of it doesn't contain anything sensitive. At the same time, the change grants access to the file to the most untrusted domains, such as untrusted_app.

Patch-set: 25
Label: Verified=0