Update patch set 25
Patch Set 25:

The history as to why this is required is in the dependent CL https://android-review.googlesource.com/#/c/273398 where we add /dev/event-log-tags. /dev/event-log-tags is a writeable file that contains runtime extensions to the readonly /system/etc/event-log-tags file.

This file by itself contains a device signature, which can be used as a way to decide on attack vectors. With the addition of /dev/event-log-tags, we will in time (as the users for the event tag service grows) also get a signature of applications on the platform. jeffv@ demanded that I protect /dev/event-log-tags at least, had to separate this into a dependent CL to simplify, and that is why this blocks c/273398.

As it stands, we can not truly protect these files (yet) because there is a bug in the sepolicies which has placed one of the system application into the untrusted_app domain. Please work on fixing this bug b/34167284.

This CL serves the purpose of stating the current affairs regarding access to /system/etc/event-log-tags, for good, or for bad. We expect that untrusted_app.te will be closed with b/34167284. We expect some followup churn as selinux violations for other application(s) trigger this, and we may have to open up, or fix, the associated code.

By itself, this is a means to test who requires the tag mapping, and should be inert, and it has already discovered one interesting access domain. By itself we might make a case for not enforcing, but then the next CL requires enforcement ...

Patch-set: 25