Update patch set 6
Patch Set 6: Code-Review-1

(1 comment)

boot_control_hal != hal_boot. hal_boot is the attribute of the service implementing the HIDL HAL which includes HIDL stuff. boot_control_hal is the attribute for the permissions needed to execute the operations defined in the device-specific HAL. The service implementing the HAL when running in normal mode will require all the stuff granted by boot_control_hal, but they are not the same.

I would prefer if you don't submit this... but that's just my opinion.

Patch-set: 6
Label: Code-Review=-1