Update patch set 1
Patch Set 1: Code-Review+1

All the HALs typically require reading the contents of directories in /{system,vendor,odm}/lib[64]/hw/. I've seen this added a bunch of time, and in fact, the ability to read the directories in /system was at one time in domain.te and granted to everyone.

This is just moving an SELinux rule from the wrong file to the correct file, but it doesn't actually change the security policies of the device. Just a cleanup, so LGTM. 

We might want to reconsider how we handle reading directories on /system at some point, or figure out a way so the HAL mechanism doesn't require directory read permissions. But I think we can do that later.

+1, but I will +2 if klyubin doesn't have anymore concerns.

Patch-set: 1