Update patch set 4
Patch Set 4:

Nick, to illustrate the macro, in PS4 I added the macro and switched hal_audio_default and a couple of more domains to it. I believe the desire behind having hal_impl_domain macro is to hide the seeming complexity of marking a domain with multiple attributes, such as haldomain and hal_audio.

Patch-set: 4