split priv_app to remove app_data_file execute
This splits priv_app based on the signing key to remove app_data_file
execute for the base system. Google Play uses different signing keys, so
the issues discovered via auditallow are addressed by running those apps
in a different domain.

The intention is to take the same approach for untrusted_app based on
signatures too, which has been confirmed to work too.

The next step would be exposing a way for third party apps to opt into
this for untrusted_app, and then ideally flipping the default for a new
API level with a way to opt out down the road. The base system is the
best way to start since it won't involve any SDK API changes.

Change-Id: Ia9e44166eaaf0cd814d510e3bff5f1e016fdc080