Update patch set 5
Patch Set 5:

I think it makes sense to have a domain for each base system app. I wanted this to be usable for other priv_apps that aren't necessarily strongly tied to the base system though. It's currently based solely on signatures, but the intent is that the same split is done for untrusted_app and then it becomes something that apps can put in their manifest to opt-in. Doing the base system via signatures is just an easy way to get this started as a sweeping change and to prevent it from regressing, since while it works right not it might not continue working if people have the option of executing app_data_file instead of shipping / updating code as signed apks where they can't modify their code (at least on storage).

Patch-set: 5