Update patch set 5
Patch Set 5:

Also if the ART JIT ever became a compile-time switch (full AOT vs. JIT or something like that - and I know that might not happen), this could gain the additional meaning of removing execmem and tmpfs execute once the WebView sandbox can be assumed to be present (maybe in a year or two?).

Patch-set: 5
Label: Verified=0