Update patch set 5
Patch Set 5:

> I think it makes sense to have a domain for each base system app.

Great!

 > I
 > wanted this to be usable for other priv_apps that aren't
 > necessarily strongly tied to the base system though.

Just like we do for services, it's reasonable to ask anyone writing a priv-app to create an SELinux domain for it. Long term, IMHO, priv-app shouldn't exist, but rather, a series of concrete SELinux domains, one for each priv-app, should exist.

This can be done using a variant of your patch. Keep priv_app_common.te, and instead of creating priv_app_nxfile.te, just create one new domain for every priv-app in /system/priv-app. And have that domain use the priv_app_common rules.

 > It's currently
 > based solely on signatures,

The seapp_contexts file can route APKs into domains based off of the signature *and packagename*. There's no need to limit this to signature alone, nor should we (IMHO).

 > but the intent is that the same split
 > is done for untrusted_app and then it becomes something that apps
 > can put in their manifest to opt-in. 

I would recommend on just focusing on priv-apps for now.

Historically, I've been reluctant to have SELinux enforce security controls based on an app supplied field. In the case where we're trying to limit access to a resource, the app can always change the field, bypassing any security control we put in place. Arguably this could be used to reduce application attack surface, but then, we'd have to produce a number of SELinux domains for different permutations of the permission. This will complicate testing and produce a huge number of domains.

If we're thinking about targeting SELinux based on something like targetAPI, I would prefer if we revive the following patches:

  https://android-review.googlesource.com/#/q/topic:selinux-targetSdkVersion+(status:open+OR+status:merged)

 > Doing the base system via
 > signatures is just an easy way to get this started as a sweeping
 > change and to prevent it from regressing, since while it works
 > right not it might not continue working if people have the option
 > of executing app_data_file instead of shipping / updating code as
 > signed apks where they can't modify their code (at least on
 > storage).

Patch-set: 5