Update patch set 6
Patch Set 6:

> Historically, I've been reluctant to have SELinux enforce security controls based on an app supplied field. In the case where we're trying to limit access to a resource, the app can always change the field, bypassing any security control we put in place. Arguably this could be used to reduce application attack surface, but then, we'd have to produce a number of SELinux domains for different permutations of the permission. This will complicate testing and produce a huge number of domains.

I agree with that, but for execmem and tmpfs/app_data_file execute the main reason to remove it is really to protect the app from itself (i.e. being exploited and then the malware persisting inside it). There are legitimate use cases for them, and it's not particularly useful as a way for a malicious app to escalate privileges. SELinux just happens to be the easiest way to control these for Android since it's around already, but there would be other ways of doing it like PaX MPROTECT and grsecurity Trusted Path Execution.

The ART JIT is the main user of execmem and tmpfs execute, but if we pretend it didn't exist or was optional, there are still application JIT compilers and Firefox uses tmpfs execute too. Chromium's JIT runs in isolated_app and the WebView can now do that (but doesn't by default), but forcing all apps above a certain API level to use isolated_app for JIT seems too aggressive in the near future. They need a lot of warning, so the first step would be forcing them to opt-in to it.

Patch-set: 6
Reviewer: Gerrit User 1069238 <1069238@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0