Update patch set 4
Patch Set 4:

Just an update on this, we're redesigning the way configfs works with ffs functions that will mean that the additional property_set permission is not needed for mediaprovider. As such there's no hurry to get this patch in, just something that would be good to have eventually.

Patch-set: 4
Reviewer: Gerrit User 1113594 <1113594@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0